+++ R/check.R
@@ -213,3 +213,28 @@ checkPlots <- function(plots) {
 
   return(NULL)
 }
+
+checkBarcodes <- function(barcodes) {
+  checkList(barcodes)
+
+  for (i in seq_along(barcodes)) {
+    barcode <- barcodes[[i]]
+    checkList(barcode)
+    statistic <- barcode[["statistic"]]
+    logFoldChange <- barcode[["logFoldChange"]]
+    absolute <- barcode[["absolute"]]
+    labelStat <- barcode[["labelStat"]]
+    labelLow <- barcode[["labelLow"]]
+    labelHigh <- barcode[["labelHigh"]]
+    stopifnot(
+      is.character(statistic),
+      length(statistic) == 1
+    )
+    elements <- setdiff(names(barcode), "statistic")
+    for (e in elements) {
+      stopifnot(length(barcode[[e]]) == 1)
+    }
+  }
+
+  return(NULL)
+}