+++ NEWS.md
@@ -4,6 +4,10 @@
 the functions `getResultsIntersection()`, `getEnrichmentsIntersection()`,
 `getResultsUpset()`, and `getEnrichmentsUpset()`.
 
+* Support customization of barcode plots:
+    * `addBarcodes()`
+    * `getBarcodes()`
+
 # 0.4.0
 
 * `getResultsIntersection()`